Add the ability to specify hash in client bundles
You can now add a [hash] placeholder inside your filename in the config. It
will be replace with the file compilation hash.

You can access to the filename / hash mapping in the browser though the global
object `__ENTRY_PATHS__`

Known caveats : If you are using a serviceWorker to cache all the resources,
it is not possible to have access to the window object. So you cannot have
the proper filenames. We are investingating on a solution. We did not find
the ideal one yet. You'll be updated when it is the case.